refactor(eppi): fold additional-text mapping into converter
- Inline eppi_output_data_from_eppi_fields branches in eppi_annotation_converter;
  remove eppi_additional_text_mapping module
- Mark lazy import noqa PLC0415 in ProcessedEppiAnnotationData (avoid circular import)

Made-with: Cursor
