+++ deet/processors/eppi_annotation_converter.py
@@ -2,7 +2,7 @@
 
 import json
 from pathlib import Path
-from typing import Any
+from typing import Any, cast
 
 from loguru import logger
 from pydantic import TypeAdapter
@@ -28,10 +28,72 @@ from deet.processors.base_converter import (
     AnnotationConverter,
     Outfiles,
 )
-from deet.processors.eppi_additional_text_mapping import (
-    EppiRawDataValue,
-    eppi_output_data_from_eppi_fields,
-)
+
+EppiRawDataValue = bool | str | int | float | list[Any] | dict[str, Any]
+
+
+def eppi_output_data_from_eppi_fields(  # noqa: PLR0911
+    output_data_type: AttributeType,
+    *,
+    additional_text: str,
+) -> EppiRawDataValue:
+    """
+    Map EPPI ``Codes`` row evidence onto typed ``raw_data`` for coerced ``output_data``.
+
+    A row in ``References[].Codes`` means the reviewer applied that code (e.g. ticked
+    the checkbox). For boolean attributes, that application is ``True`` even when
+    ``AdditionalText`` is empty.
+
+    For every non-boolean type, only the info-box ``AdditionalText`` is used.
+    ``ItemAttributeFullTextDetails`` is not used for the stored value (it may still be
+    attached to the model for other uses).
+
+    Args:
+        output_data_type: Target attribute type (from codeset or prompt CSV).
+        additional_text: EPPI ``AdditionalText`` / info-box value.
+
+    Returns:
+        Value to store in ``GoldStandardAnnotation.raw_data`` (then coerced via
+        ``output_data``).
+
+    """
+    additional = (additional_text or "").strip()
+
+    if output_data_type == AttributeType.BOOL:
+        return True
+    if output_data_type == AttributeType.STRING:
+        return additional
+
+    if output_data_type == AttributeType.INTEGER:
+        if not additional:
+            return output_data_type.missing_annotation_default()
+        try:
+            return int(float(additional))
+        except ValueError:
+            return output_data_type.missing_annotation_default()
+
+    if output_data_type == AttributeType.FLOAT:
+        if not additional:
+            return output_data_type.missing_annotation_default()
+        try:
+            return float(additional.replace(",", ""))
+        except ValueError:
+            return output_data_type.missing_annotation_default()
+
+    if output_data_type in (AttributeType.LIST, AttributeType.DICT):
+        if not additional:
+            return output_data_type.missing_annotation_default()
+        try:
+            parsed: Any = json.loads(additional)
+        except (json.JSONDecodeError, TypeError):
+            return output_data_type.missing_annotation_default()
+        py_type = output_data_type.to_python_type()
+        if isinstance(parsed, py_type):
+            return cast("EppiRawDataValue", parsed)
+        return output_data_type.missing_annotation_default()
+
+    unsupported = f"Unsupported AttributeType for EPPI mapping: {output_data_type}"
+    raise ValueError(unsupported)
 
 
 class EppiAnnotationConverter(AnnotationConverter):
@@ -257,8 +319,10 @@ class EppiAnnotationConverter(AnnotationConverter):
             attribute.attribute_label = attribute_id_to_label[attribute_id]
 
         additional_text = str(annotation.get("AdditionalText", "") or "")
-        typed_raw_data: EppiRawDataValue = eppi_output_data_from_eppi_fields(
-            attribute.output_data_type, additional_text=additional_text
+        typed_raw_data: bool | str | int | float | list[Any] | dict[str, Any] = (
+            eppi_output_data_from_eppi_fields(
+                attribute.output_data_type, additional_text=additional_text
+            )
         )
 
         return EppiGoldStandardAnnotation(
